requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
websockets==12.0
pytest==7.4.4
pytest-asyncio==0.23.3